move test outside of __global__
